Split medicaid variables into own files
