+++ openfisca_us/variables/hhs/medicaid/medicaid_person_type.py
@@ -1,0 +1,37 @@
+from openfisca_us.model_api import *
+
+
+class MedicaidPersonType(Enum):
+    ADULT_WITHOUT_DEPENDENT = "Adult without dependent"
+    ADULT_WITH_DEPENDENT = "Adult with dependent"
+    CHILD_AGE_0 = "Child age 0"
+    CHILD_AGE_1_5 = "Child age 1 to 5"
+    CHILD_AGE_6_18 = "Child age 6 to 18"
+
+
+class medicaid_person_type(Variable):
+    value_type = Enum
+    possible_values = MedicaidPersonType
+    default_value = MedicaidPersonType.ADULT_WITHOUT_DEPENDENT
+    entity = Person
+    definition_period = YEAR
+    label = "Medicaid person type"
+    documentation = "Person type for Medicaid"
+
+    def formula(person, period, parameters):
+        # Get the person's age.
+        age = person("age", period)
+        # Get the existence of dependents, as defined by people 18 or younger.
+        has_dependents = person.spm_unit.any(
+            person.spm_unit.members("age", period) < 19
+        )
+        return select(
+            [age == 0, age < 6, age < 19, has_dependents, True],
+            [
+                MedicaidPersonType.CHILD_AGE_0,
+                MedicaidPersonType.CHILD_AGE_1_5,
+                MedicaidPersonType.CHILD_AGE_6_18,
+                MedicaidPersonType.ADULT_WITH_DEPENDENT,
+                MedicaidPersonType.ADULT_WITHOUT_DEPENDENT,
+            ],
+        )
